requirements.txt
streamlit
google-generativeai
gTTS
requests
Pillow
